perf(core): separate the cache file
