+++ internal/core/cache.go
@@ -6,9 +6,7 @@ import (
 	"os"
 	"os/user"
 	"path"
-	// "strings"
 
-	"github.com/darwinia-network/shadow/internal"
 	"github.com/darwinia-network/shadow/internal/eth"
 	"github.com/darwinia-network/shadow/internal/log"
 	"github.com/darwinia-network/shadow/internal/util"
@@ -19,245 +17,6 @@ import (
 
 // Same directory as `darwinia.js`
 const DB_PATH = ".darwinia/cache/shadow.db"
-
-// EthHeaderWithProof Cache
-type EthHeaderWithProofCache struct {
-	Hash   string `json:"hash"`
-	Number uint64 `json:"number" gorm:"unique_index"`
-	Header string `json:"eth_header"`
-	Proof  string `json:"proof"`
-	Root   string `json:"root" gorm:"DEFAULT:NULL"`
-}
-
-func CreateEthHeaderCache(db *gorm.DB, header types.Header) error {
-	if util.IsEmpty(header) {
-		return fmt.Errorf("empty header")
-	}
-
-	dh, err := eth.IntoDarwiniaEthHeader(header)
-	if err != nil {
-		return err
-	}
-
-	hstr, err := dh.ToString()
-	if err != nil {
-		return err
-	}
-
-	cache := EthHeaderWithProofCache{
-		Hash:   header.Hash().String(),
-		Number: header.Number.Uint64(),
-		Header: hstr,
-	}
-
-	sql := fmt.Sprintf(
-		"%s %s %s %s ('%s', %v, '%s');",
-		"INSERT or REPLACE INTO",       // insert or replace
-		"eth_header_with_proof_caches", // table
-		"(hash, number, header)",       // columns
-		"values",                       // values
-		cache.Hash,
-		cache.Number,
-		cache.Header,
-	)
-
-	if err := db.Exec(sql).Error; err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func CountCache(db *gorm.DB) uint64 {
-	var count uint64
-	db.Model(&EthHeaderWithProofCache{}).Count(&count)
-	return count
-}
-
-// Save header to cache
-func (c *EthHeaderWithProofCache) FromResp(
-	db *gorm.DB,
-	resp GetEthHeaderWithProofRawResp,
-) error {
-	// Convert header to string
-	header, err := json.Marshal(resp.Header)
-	if err != nil {
-		return err
-	}
-
-	proof, err := json.Marshal(resp.Proof)
-	if err != nil {
-		return err
-	}
-
-	db.Create(&EthHeaderWithProofCache{
-		Hash:   resp.Header.Hash,
-		Number: resp.Header.Number,
-		Header: string(header),
-		Proof:  string(proof),
-	})
-
-	return nil
-}
-
-/// The func should run after `Fetch`
-func (c *EthHeaderWithProofCache) ApplyProof(
-	config internal.Config,
-	db *gorm.DB,
-) error {
-	var (
-		ethHeader types.Header
-		err       error
-	)
-
-	if util.IsEmpty(c.Number) && c.Number != 0 {
-		log.Warn("Apply ethashproof to block %v failed", c.Number)
-		return fmt.Errorf("Apply ethashproof to block %v failed", c.Number)
-	} else if util.IsEmpty(c.Header) || c.Header == "" {
-		log.Warn("Can not analyse the block %v", c)
-		return fmt.Errorf("Can not analyse the block %v", c)
-	}
-
-	log.Trace("Apply ethashproof to block %v...", c.Number)
-	// Check proof lock
-	if util.IsEmpty(c.Proof) || c.Proof == "" {
-		if config.CheckLock(PROOF_LOCK) {
-			return fmt.Errorf("Shadow service is busy now, please try again later")
-		} else {
-			err := config.CreateLock(PROOF_LOCK, []byte(""))
-			if err != nil {
-				return err
-			}
-		}
-
-		// Fetch EthHeader
-		ethHeader, err = eth.Header(c.Number, config.Api)
-		if err != nil {
-			return err
-		}
-
-		// Proof header
-		proof, err := eth.Proof(&ethHeader, config)
-		if err != nil {
-			return err
-		}
-
-		proofBytes, err := json.Marshal(proof.Format())
-		if err != nil {
-			return err
-		}
-
-		c.Proof = string(proofBytes)
-		err = db.Model(&c).Where("number = ?", c.Number).Update("proof", c.Proof).Error
-		if err != nil {
-			return err
-		}
-
-		// Remove proof lock
-		err = config.RemoveLock(PROOF_LOCK)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// Convert EthHeader
-func (c *EthHeaderWithProofCache) IntoResp() (GetEthHeaderWithProofRawResp, error) {
-	var rResp GetEthHeaderWithProofRawResp
-	header, proof := eth.DarwiniaEthHeader{}, []eth.DoubleNodeWithMerkleProof{}
-
-	// Decode header
-	err := json.Unmarshal([]byte(c.Header), &header)
-	if err != nil {
-		log.Error("Unmarshal eth header failed %v", c.Header)
-		return rResp, err
-	}
-
-	// Decode proof
-	if !util.IsEmpty(c.Proof) {
-		err = json.Unmarshal([]byte(c.Proof), &proof)
-		if err != nil {
-			log.Error("Unmarshal eth header proof failed %v", c.Proof)
-			return rResp, err
-		}
-	}
-
-	// Construct resp
-	return GetEthHeaderWithProofRawResp{
-		Header: header,
-		Proof:  proof,
-		Root:   c.Root,
-	}, nil
-}
-
-// Fetch Eth Header by number
-func (c *EthHeaderWithProofCache) Fetch(
-	config internal.Config,
-	db *gorm.DB,
-	geth *eth.Geth,
-) error {
-	// Get header from sqlite3
-	var (
-		err   error
-		block interface{}
-	)
-
-	if !util.IsEmpty(c.Number) || c.Number == 0 {
-		block = c.Number
-		err = db.Where("number = ?", c.Number).Take(&c).Error
-	} else if !util.IsEmpty(c.Hash) {
-		block = c.Hash
-		err = db.Where("hash = ?", c.Hash).Take(&c).Error
-	}
-
-	if err != nil || util.IsEmpty(c.Header) || c.Header == "" {
-		var ethHeader types.Header
-
-		// Log the event
-		log.Trace("fetching block %v ...", block)
-		if !util.IsEmpty(geth) {
-			block := *geth.Header(block)
-			if !util.IsEmpty(block) {
-				ethHeader = block
-			}
-		}
-
-		if util.IsEmpty(ethHeader) {
-			ethHeader, err = eth.Header(block, config.Api)
-			if err != nil {
-				return err
-			}
-		}
-
-		header, err := eth.IntoDarwiniaEthHeader(ethHeader)
-		if err != nil {
-			return err
-		}
-
-		bytes, err := json.Marshal(header)
-		if err != nil {
-			return err
-		}
-
-		c.Header = string(bytes)
-		c.Hash = ethHeader.Hash().Hex()
-		c.Number = header.Number
-		db.Create(&c)
-
-		// Prints logs every 100 headers
-		if c.Number > 0 && c.Number%100 == 0 {
-			log.Info(
-				"imported headers from #%v to #%v\n",
-				c.Number-100,
-				c.Number,
-			)
-		}
-	}
-
-	// Return resp
-	return nil
-}
 
 // Connect to cache
 func ConnectDb() (*gorm.DB, error) {
@@ -287,4 +46,107 @@ func ConnectDb() (*gorm.DB, error) {
 	// bootstrap sqlite3
 	db.AutoMigrate(&EthHeaderWithProofCache{})
 	return db, err
+}
+
+func CountCache(db *gorm.DB) uint64 {
+	var count uint64
+	db.Model(&EthHeaderWithProofCache{}).Count(&count)
+	return count
+}
+
+func FetchHeader(shadow *Shadow, block interface{}) (types.Header, error) {
+	num, err := shadow.checkGenesis(shadow.Config.Genesis, block)
+	if err != nil {
+		return types.Header{}, err
+	}
+
+	if !util.IsEmpty(shadow.Geth) {
+		block := *shadow.Geth.Header(block)
+		if !util.IsEmpty(block) {
+			return block, nil
+		}
+	}
+
+	return eth.Header(num, shadow.Config.Api)
+}
+
+func CreateEthHeaderCache(db *gorm.DB, header types.Header) error {
+	if util.IsEmpty(header) {
+		return fmt.Errorf("empty header")
+	}
+
+	dh, err := eth.IntoDarwiniaEthHeader(header)
+	if err != nil {
+		return err
+	}
+
+	hstr, err := dh.ToString()
+	if err != nil {
+		return err
+	}
+
+	cache := EthHeaderWithProofCache{
+		Hash:   header.Hash().Hex(),
+		Number: header.Number.Uint64(),
+		Header: hstr,
+	}
+
+	if err := db.Exec(UpsertHeaderSQL(&cache)).Error; err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func CreateProofCache(
+	shadow *Shadow,
+	cache *EthHeaderWithProofCache,
+	header *types.Header,
+) error {
+	// Proof header
+	proof, err := eth.Proof(header, shadow.Config)
+	if err != nil {
+		return err
+	}
+
+	proofBytes, err := json.Marshal(proof.Format())
+	if err != nil {
+		return err
+	}
+
+	cache.Proof = string(proofBytes)
+	err = shadow.DB.Model(&cache).Where(
+		"number = ?", cache.Number,
+	).Update(
+		"proof", cache.Proof,
+	).Error
+
+	if err != nil {
+		return err
+	}
+
+	// Remove proof lock
+	err = shadow.Config.RemoveLock(PROOF_LOCK)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func IndexHeaderFromDB(db *gorm.DB, cache *EthHeaderWithProofCache) (interface{}, error) {
+	var (
+		err   error
+		block interface{}
+	)
+
+	if !util.IsEmpty(cache.Number) || cache.Number == 0 {
+		block = cache.Number
+		err = db.Where("number = ?", cache.Number).Take(&cache).Error
+	} else if !util.IsEmpty(cache.Hash) {
+		block = cache.Hash
+		err = db.Where("hash = ?", cache.Hash).Take(&cache).Error
+	}
+
+	return block, err
 }
